fix(image handling): Add image_url to JSON data in relevant CBV for artworks list
Refactored the sort buttons on artwork_list.html and it's relevant view to handle sorting via SSR

In manual testing found that the SSR and JSON progressive enhancement weren't handling the images for artworks list properly. Added image_url to JSON data to enable it to work in dev environment as well as staging and production.

Documented fixes and credits in the README.md

Files Changes:
-  README.md
- pointless_impressions_src/artwork/*
- pointless_impressions_src/theme/static_src/*

Still to do:
- Check the US001 features pass the manual tests